refactor(seeder): Apply naming changes suggested in PR review
